moved utils to global, moved time arrays to global, fixed routes
