+++ src/InterviewQ/components/History/subs/CoachHistory/1_CoachHistory.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useQuery } from '@apollo/react-hooks';
 import { gql } from 'apollo-boost';
 
-import { isPast } from '../../../../../utils/isPast';
+import { isPast } from '../../../../../global/utils/isPast';
 
 import CoachHistoryRow from './2_CoachHistoryRow';
