+++ src/global/components/Calendar/SmallCalendar.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import '../Dashboard/Schedule/Calendar.scss';
+import '../Dashboard/subs/Schedule/Calendar.scss';
 import { setMonth, getMonth, getYear, addMonths, subMonths } from 'date-fns';
 
 import { nextArrow } from '../../icons/nextArrow';
@@ -7,7 +7,7 @@ import { backArrow } from '../../icons/backArrow';
 
 import SmallCells from './SmallCells';
 
-import { days, months, years } from '../Dashboard/Schedule/TimeArrays'
+import { days, months, years } from '../../utils/TimeArrays';
 
 const SmallCalendar = ({ selectedCell, setSelectedCell, availabilities, refetchAvails }) => {
 	const [currentMonth, setCurrentMonth] = useState(new Date());
