+++ src/global/components/Dashboard/Dashboard.js
@@ -10,7 +10,7 @@ import './Dashboard.scss';
 // Components
 import LeftNavBar from './subs/LeftNavBar';
 import PersonalInfo from './subs/PersonalInfo/PersonalInfo';
-import Schedule from './Schedule';
+import Schedule from './subs/Schedule';
 // import DashInterviewQ from '../../../InterviewQ/components/DashInterviewQ';
 // GraphQuaiL Query
 const GET_USER = gql`
