+++ src/global/components/Dashboard/subs/Schedule/Schedule.js
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import { Route } from 'react-router-dom';
 import Calendar from './Calendar.js'
 import WeekView from './WeekView.js';
-import Icon from '../../../icons/Icon';
-import { ICONS } from '../../../icons/iconConstants';
+import Icon from '../../../../icons/Icon';
+import { ICONS } from '../../../../icons/iconConstants';
 
 const Schedule = () => {
 	const [selectedDate, setSelectedDate] = useState(new Date());
