+++ src/global/utils/capitalize.js
